# docs/swagger/index.yml
openapi: 3.0.3
info:
  title: online-shop-express
  version: 1.0.0
  contact: {}
servers:
  - url: 127.0.0.1
paths:
...
  /api/v1/orders:
    get:
      tags:
        - order-service
      summary: 'check-orders '
      description: 'check-orders '
      operationId: checkOrders
      parameters:
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: size
          in: query
          schema:
            type: string
            example: '2'
      requestBody:
        content:
          application/json:
            examples:
              'check-orders ':
                value: ''
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/order/656dabab38cbcee66c3a240a:
    get:
      tags:
        - order-service
      summary: read-order-by-id
      description: read-order-by-id
      operationId: readOrderById
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...